+++ CHANGELOG.md
@@ -10,6 +10,7 @@ A preview of the next release can be installed from
 ## Unreleased
 
 - [#1507](https://github.com/babashka/babashka/issues/1507): Expose methods on java.lang.VirtualThread ([@lispyclouds](https://github.com/lispyclouds))
+- [#1510](https://github.com/babashka/babashka/issues/1510): add virtual thread interop on `Thread`
 
 ## 1.2.174 (2023-03-01)